Move navigator tests into a separate file.
